requirements.txt
pydantic==2.5.2
streamlit==1.31.0
requests==2.31.0
sqlite3-api==2.0.4
typing-extensions==4.9.0
python-dotenv==1.0.0